requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
Pillow==10.0.1
pydantic==2.4.2
requests==2.31.0